Puzzle refactoring
